requirements.txt
async-generator==1.10
attrs==21.4.0
Automat==20.2.0
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
h11==0.13.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
Jinja2==3.1.1
jmespath==0.10.0
lxml==4.7.1
MarkupSafe==2.1.1
outcome==1.1.0
parsel==1.6.0
Pillow==9.0.1
pkg_resources==0.0.0
priority==1.3.0
Protego==0.1.16
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==21.0.0
python-dotenv==0.20.0
PyYAML==6.0
queuelib==1.6.2
Scrapy==2.5.1
selenium==4.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
toml==0.10.2
trio==0.19.0
trio-websocket==0.9.2
Twisted==21.7.0
typing_extensions==4.0.1
urllib3==1.26.8
w3lib==1.22.0
wsproto==1.0.0
zope.interface==5.4.0
